CapstoneProject-Spoken_claim_verification/requirements.txt
# Core LLM and API Libraries
openai==1.3.0
google-search-results==2.4.2
google-generativeai==0.3.0

# Speech Recognition
openai-whisper==20231117
librosa==0.10.0

# Data Processing
pandas==2.0.3
numpy==1.24.3

# Database
mysql-connector-python==8.0.33
influxdb-client==1.18.0
sqlalchemy==2.0.19

# Web Framework
streamlit==1.28.0
fastapi==0.104.1
uvicorn==0.24.0

# Data Visualization
plotly==5.17.0
matplotlib==3.7.2
seaborn==0.12.2

# Utilities
python-dotenv==1.0.0
pyyaml==6.0
requests==2.31.0
tqdm==4.66.1

# Audio Processing
pydub==0.25.1
scipy==1.11.2

# JSON and Data Validation
pydantic==2.3.0
jsonschema==4.19.0

# Logging and Monitoring
python-json-logger==2.0.7

# Testing
pytest==7.4.2
pytest-cov==4.1.0
